+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to core_tools will be documented in this file.
 - Lowered VideoMode refresh rate to better support remote connections
 - Added option silent to measurement and scan_generic
 - Keyboard interrupt in measurement now aborts running script (So you don't have to hit Ctrl-C 50 times..)
+- Added get_idn to Instruments gates, hardware, M3102A.
 
 ## \[1.4.6] - 2022-10-17
 - Improved logfile format