+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to core_tools will be documented in this file.
 - Added option silent to measurement and scan_generic
 - Keyboard interrupt in measurement now aborts running script (So you don't have to hit Ctrl-C 50 times..)
 - Added get_idn to Instruments gates, hardware, M3102A.
+- Fixed resetting histogram
 
 ## \[1.4.6] - 2022-10-17
 - Improved logfile format