+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to core_tools will be documented in this file.
 - Keyboard interrupt in measurement now aborts running script (So you don't have to hit Ctrl-C 50 times..)
 - Added get_idn to Instruments gates, hardware, M3102A.
 - Fixed resetting histogram
+- Added option to disable live plotting of data browser at startup
 
 ## \[1.4.6] - 2022-10-17
 - Improved logfile format